refactor : core와 storage 디렉토리 분리
