+++ src/main/java/com/pintoss/auth/core/payment/application/PaymentRepository.java
@@ -1,6 +1,6 @@
 package com.pintoss.auth.core.payment.application;
 
-import com.pintoss.auth.core.payment.store.PaymentEntity;
+import com.pintoss.auth.storage.payment.PaymentEntity;
 
 public interface PaymentRepository {
