+++ src/main/java/com/pintoss/auth/storage/order/OrderJpaRepository.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.order.store;
+package com.pintoss.auth.storage.order;
 
 import com.pintoss.auth.core.order.domain.Order;
 import java.util.Optional;
