+++ src/main/java/com/pintoss/auth/storage/voucher/VoucherIssuerJpaRepository.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.voucher.store;
+package com.pintoss.auth.storage.voucher;
 
 import org.springframework.data.jpa.repository.JpaRepository;
