+++ src/main/java/com/pintoss/auth/storage/voucher/VoucherQueryDslRepository.java
@@ -1,8 +1,6 @@
-package com.pintoss.auth.core.voucher.store;
+package com.pintoss.auth.storage.voucher;
 
 
-import static com.pintoss.auth.core.voucher.store.QVoucherEntity.voucherEntity;
-import static com.pintoss.auth.core.voucher.store.QVoucherIssuerEntity.voucherIssuerEntity;
 
 import com.pintoss.auth.core.voucher.domain.Voucher;
 import com.querydsl.core.types.Projections;
@@ -10,6 +8,9 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+
+import static com.pintoss.auth.storage.voucher.QVoucherEntity.voucherEntity;
+import static com.pintoss.auth.storage.voucher.QVoucherIssuerEntity.voucherIssuerEntity;
 
 @Repository
 @RequiredArgsConstructor
